Fix line numbers reported for standalone iterators
As originally mentioned in 189feea8f0ab5c2c94dac26b3d317a68246b14a9 we were
reporting the wrong line numbers for order independent loops that invoked a
standalone iterator. Instead of reporting the line number of the user-level
loop that was order independent, we were reporting the line number of the
yielding loop inside the standalone iterator. This was particularly annoying
for the forallInStandalone test because it meant code changes that affected
ChapelRange line numbers would cause it to fail.

This fix is peeled off of (blatantly stolen from) a larger patch that Vass
made. His patch updates when we preserve line numbers. This ignores that
question, and sticks with our current line number reporting scheme, but
corrects what we report for invocations of standalone iterators.