Fix bugs in fifo and clean up; fix execution order problems in two tests.
The big change here is a clean-up in thread creation.  We try to run as
many threads as we have active+queued tasks, so that no tasks have to
wait for a thread.  This was handled by launch_next_task_in_new_thread()
and schedule_next_task().  As an optimization, when we created a new
thread we gave it its first task to run directly; it didn't have to
search for it in the pool.  I decided this probably isn't really an
optimization.  Creating a thread takes a long time (ultimately it's a
pthread_create() kernel call).  If we happen to have an idle thread
looking for tasks already, it's way faster to let that one pick up a
newly queued task than to force the task to wait until a newly-created
thread is running.  Also, although the two functions mentioned were
coded to start more than one thread at a time, in practice we only used
them to start one.  So the changes are:
- Merge the code in the two functions above into the new function
  maybe_add_thread().  This starts a thread if it can, but it doesn't
  feed that thread a task.  The new thread can look in the pool itself
  for a task to run, after it starts.
- In thread_begin(), rework (rotate) the loop that finds and runs tasks,
  to reflect that a new thread doesn't start with an assigned task but
  instead has to find work on its own.  I also cleaned up the loop in
  other ways -- the logic for searching for a task and suspending when
  there wasn't anything to do was unnecessarily complicated.

In enqueue_task() and dequeue_task(), I fixed a problem wherein list
maintainance was actually manipulating a link belonging to the pool, not
the list.  I suspect this was the cause of the segfaults and other
intermittent problems I was seeing before.  I also moved maintenance of
the queued task count into enqueue_task() and dequeue_task().

I got rid of the "begun" flag, because now a task cannot be running
without first having been removed from the pool anyway.  I also got rid
of waking_thread_cnt because if threads don't start with tasks already
assigned we don't need to count waking threads specially.

For the test/parallel/cobegin/stonea/reports test, I adjusted the
prediff and the good to ignore the source line numbers of the tasks.
This means we still check for the right number of active and pending
tasks, but it doesn't matter which tasks are active or pending.  That
is, we now ignore variation in task execution order by the tasking
layer.

Finally, I changed the test/parallel/taskPool/figueroa/QueuedTasks test
so that it doesn't require the tasking layer to run the begin-stmt tasks
in a specify order or deadlock.  It can now run them in either order.