Split partially bounded range builders out for low and high bounded ranges
Split `chpl_build_partially_bounded_range` into `chpl_build_low_bounded_range`
and `chpl_build_high_bounded_range`. This is in preparation for an optimization
of anonymous counted low bounded ranges.

Since the bound type is now embedded in the name, the parser doesn't need to
pass the BoundRangeType to the builder. Splitting out the builders also allows
us to clean up/improve the error messages, builder signature, and the calls to
the range constructors.