Fix ISx sed portability issue
Replace ISx `sed -i`, which isn't portable across the platforms we test on
with a more generic and portable method.  Traditionally, we just do a regular
sed, send the output to a temp file, and then replace the original with the
temp file. Do that here in a loop over the .good files.