Allow performance graphs to be generated concurrently
We used to generate performance graphs on a per machine basis so there was
never any way for genGraphs to be running concurrently on the same directory.
However, we started running multiple performance tests on chapcs boxes and they
all generated graphs in the chapcs/html directory instead of the machine
specific directory.

This makes us build in chapcs/<performancedescription>/html so we have a
unique directory now.