Protect against compiler segfault
This segfault previously occurred when multiple nested modules with the same
name defined a main function and the ambiguous name was specified via the
--main-module flag.  DeadCodeElimination was expecting a value that it could
call a method on, but the value it was given was NULL, likely due to an earlier
bug in the compiler.  This change just causes us to get an internal error
message instead, the actual fix will be more involved.