Extend PR #1351 to apply to 'align' as well and add comments
In PR #1351, I started renaming 'by' to 'chpl_by' in the front-end
to avoid preventing the use of 'by' as an identifer in roundabout
ways (such as having a file named by.chpl which would introduce an
identifier named 'by' to describe the implicit module by default.

This commit extends that support to include 'align' as well which
falls into precisely the same category (overloadable operators
that are otherwise legal identifiers).  It also adds comments to
help understand that 'proc by' overloads in the module code are
effectively equivalent to 'proc chpl_by' to aid with those who
might be grepping for the definition of said routine.