Added some comments to new module code; updated tests
For the most part, any tests that tested bad sync/single types
I unified into a single testBadSyncSingle.chpl test and
commented out their code that exercised those bad types (usually
they exercised good types as well).

The test record3 I made into an 'error message' future because it
reveals that the compiler error is reported in terms of an internal
line number rather than user code.  As I noted in the .future file, I
think it's better for us to be reporting an error for a case that we
don't want to support with a bad code location than it is for us to
continue to let these sync cases slip by.