Repair the type of a ref to a "task list".
In #3178 I changed the type of the "task list" argument to the tasking
interface functions for structured parallelism.  Unfortunately in one
spot having to do with that type, I dropped a "reference to" qualifier.
I had meant to check with someone about that code (being unfamiliar with
the compiler), but forgot to do so.