Add work to cobegin-stacksize.
This test began failing with CHPL_TASKS=fifo after #3178 yesterday.  To
be specific, with the smaller 16k stack size it ran to completion rather
than segfaulting as expected.  It's not clear whether this is because
the stack requirements per recursive iteration have become smaller, or
because better task launch performance meant that the thread hosting the
parent task got to both child tasks before a new thread could do so (and
somehow that thread has more stack to work with?).  I wasn't able to get
the test to fail (that is, run to completion) manually, so I'm inclined
to think the latter explanation is the case, meaning that the failure
was sporadic/intermittent.

Thus, in this commit we bump up the depth of recursion from 1024 to
10000.  I've checked on x86_64 and recursive_sum() takes 64 bytes per
invocation there, so a depth of 10000 should be definitely not doable
with a stacksize of 16k, and doable with a stacksize of 2m.

This test is arguably a fragile way to test what it purports to test,
but we haven't come up with anything better.