Refactor some of the changes to scopeResolve into methods on the UseStmt class
Reimplements the storage of an 'except' or 'only' list into a single vector
and a flag.  Add a constructor which takes the 'only' or 'except' list, update
UseStmt's copyInner and verify methods, and creates two methods which handle
some of scopeResolve's new functionality - validateList, which ensures that
every name in the 'only' or 'except' list is a capable of being found in the
scope being searched; and createRelatedNames, which takes the resolved symbol
and if it is a type, adds the type's methods and fields to the relatedNames
list so that we don't end up with access to a type but not its inner workings.
These methods are practically identical to the functions I introduced in an
earlier attempt at this change, except they take advantage of class inner
knowledge and improve the abstraction.