Implement some helpers and allow 'only'/'except' to work on functions
Add three helper functions and an interface function for use in scopeResolve
and functionResolution.  isPlainUse returns true if the use was not created
with an 'only' or 'except' list. matchedNameOrConstructor determines if the
name provided to it is in the first level list, either on its own or because it
is a (type) constructor on a type that is in the list.  inRelatedNames
determines if the name provided to it is in the second level list, i.e. if the
name being searched for is a method or field on a type in the 'only' or 'except'
list.  skipSymbolSearch, the interface method, makes use of all of these helpers
to determine if the name provided to it is something that could be found by
looking at this use.

functionResolution was modified to call skipSymbolSearch when looking for
visible functions to match.  The modification to scopeResolve is slightly
trickier.