Adjust cobegin-stacksize behavior again.
Despite yesterday's work, this test failed again last night when run
with valgrind in a CHPL_TASKS=fifo configuration.  My guess that the
special valgrind pthread handling delayed the arrival of the host
pthreads enough that the process itself did both tasks in the second
cobegin, and for some reason the stack limit didn't apply.  (It's
supposed to be, there's code to set the limit in the process, but I
suspect there may be circumstances in which doing so isn't actually
effective.)

In any case, in this commit we try a little harder to get actual
pthreads (with the stack limits set) to run the recursive_sum() calls.