Reduce internal module line numbers reported in errors.
COMPILER CHANGES

* Factor out "resolveCall(call); resolveFns(call->isResolved())"
into resolveCallAndCallee(call).

* Within resolveCallAndCallee(), push 'call' onto 'callStack'
before resolveFns(), then pop it after resolveFns().

* Added INT_ASSERT(call->isResolved()) in there
because a couple of (former) callers of resolveFns() did that.

* In one case, removed the test if (call->isResolved())
before resolveFns(), because the test always passes.

EXPLANATION

The tests testBadSyncSingle.chpl and record3.chpl (see below)
caused the compiler to execute resolveCall+resolveFns
within resolvePrimInit(), on a call to _defaultOf().
Within _defaultOf(), there happens to be a compilerError
saying that you can't have a syncvar over type XXX.
Since that's in an internal module, the compiler goes up 'callStack'
to find a non-internal-module.

Since the call to _defaultOf was not put on callStack, the compiler
picks the bottom of the stack, which is an internal-module call
in record3 and a call to main() in testBadSyncSingle.
Neither is desirable.

Putting the call to _defaultOf() onto 'callStack' lets the compiler
pick that call, which is at the line number where that syncvar is
being declared. So we see the desired line number in the error message.

TEST CHANGES

* This test now reports adequate line numbers, so I adjusted .good files:

    parallel/sync/bradc/testBadSyncSingle

* This .future test now reports adequate line numbers as desired,
  so I un-futurized it:

    test/parallel/sync/waynew/record3