Avoid triggering ref pair when one function is a promotion
Without this rule, the combination of the normal this() array accessor
overloads and the this() methods taking in a range can make a ref pair,
where one of them uses promotion. The problem is visible with
a test using an array slice.

(However... array slicing is still not working correctly for some
 reason... )