Fix labelprop-tweets baseline regression
The recently add labelprop-tweets tests was regressing with baseline because
the refType wasn't set for a tmp in `replaceLocalWithFieldTemp()`, leading to a
failed INT_ASSERT(). This just replaces a use of `<type>.retType` with
`<type>.getRefType()`. getRefType() includes logic to check if the current type
is a ref or wide ref, before checking `<type>.retType`