Last of the interesting work refactored and a test update
Add call to skipSymbolSearch method to scopeResolve (similar to insertion in
functionResolution).  Also adds several functions/methods to support scope
resolution.  The function skipUse is provided to determine whether a particular
use statement has already been covered by other uses of that module - it does
this by relying on a new method on UseStmts, providesNewSymbols, which takes in
another UseStmt and determines whether it allows symbols to be seen in that
module which were hidden by the other use.  providesNewSymbols has some
complicated logic in it.  skipUse is utilized by buildBreadthFirstModuleList.
That function also makes use of the method applyOuterUse, which takes a use
statement that calls the parent module of this use and determines how that outer
use affects the visibility of symbols in this inner on - for instance, if the
inner use contained an 'only' list and one of the symbols in that list was in
the outer use's 'except' list, we'd want a modified version of the inner use
that wouldn't include that symbol, since in this context we don't want to see
it.

The line numbers on the expected .good for a test were wrong, so I fixed that,
too.