Implementation of 'except' and 'only' list for the AstTraversal classes
Allow the traversal classes to go through the 'except' or 'only' list.  For
code reuse, I made a helper to print out the predicate depending on which type
of list it was.  Needed to change my mind on whether certain fields and methods
were private or public to allow this to work.  Ideally, our traversal classes
wouldn't require this of me, but that's the world we live in.