Convert _do_format to take a string and rename to chpl_do_format
Noticed the casting/coercing back and forth from string->c_string->string
that's going on here and removed the c_string middle-man.  Checking in
with Michael, he believed that the use of the c_string was antiquated and
used to try and avoid leaks (in any case, he agreed that string was more
appropriate).

Interpreted this as a "meant-for-internal-use" feature, so tacked a
chpl_ prefix on it while here (the current favorite way to "guard"
an internal symbol.