Updating error message to include halt header
I'm not sure how this was getting printed without the halt header
previously...  It seems to be the only place where the given message
is printed.  Maybe it was some sort of order of initialization issue
when we were supporting c_string -> string coercions?  In any case,
this seems correct now.