Updating tests relative to removing c_string -> string coercions
These tests were relying on c_string -> string coercions, which
aren't there anymore, so adding explicit casts as well.

The c_string tests here seem to have less and less purpose in life
and could probably be removed, but it was easier for me to update
them than to determine that.