Retire tupleOfTupleIteration future
This test was getting a segfault in `replaceLocalWithFieldTemp()` when it was
attempting to get the refType of a type that was already a ref. #3212 replaceed
the `<type>.refType` with a `<type>.getRefType()` which first checks if the
current type is a ref, avoiding the null pointer that tupleOfTupleIteration was
getting.