Save the perfSha to the performance description subdir
We've been getting sporadic failures to geneate the perSha file for nightly
chapcs performance testing. This was because we were generating the file to
chapcs/perfsha.dat instead of chapcs/<performancedescription>/perfsha.dat.
This led to a race between all the chapcs perf runs, occasionally resulting
in one deleting the perfsha.dat file, while another one was trying to use it.