Fix a bug I discovered when using a standard module with 'only'/'except' lists
The bug was that if one of these lists was attached to the use, we'd fail to
find the module period.  Turns out I'd missed adding the module to our parse
list with these special uses, and hadn't seen it because user modules got added
just fine and I forgot to test uses of standard modules.  I discovered my
mistake by a little trick Brad suggested I try to see if there were good places
to insert 'only' or 'except' lists, and when I tried to add the lists, I
encountered this error.

Add three tests of this behavior - one for only, one for except, and one for
only when used inside a function scope (just in case!).