Bug fix: make PRIM_IS_SUBTYPE handle transitive instantiation
Retires test/users/ferguson/generic-default-record.future

The problem was that PRIM_IS_SUBTYPE only checked one level of
instantiation. Arguably it should have a different name, but
PRIM_IS_SUBTYPE is what a : expression maps to in a where clause. Even
before this PR, it would return true if lt is an instantiation of rt.

In this PR, I adjust it to transitively check instantiation and
sub-typing. For some reason, in the particuar generic-default-record
example, the record R had two levels of instantiation.