(Was:) Add four tests of the renaming capability.
One verifies that renaming is invalid in an 'except' list (it is useless to
rename a symbol that you aren't even going to use).  One verifies the base
case of renaming in an 'only' list.  The other two verify that you can rename
multiple symbols in an 'only' list and both will be valid (in the case where
those are the only symbols in the module, and the case where there are
additional symbols).

This is a fix up commit.  For some reason, how I did 'git cherry-pick' dropped
this first commit on the floor.  Repairing the damage.

Conflicts:
	test/visibility/rename/renameExceptInvalid.chpl
	test/visibility/rename/renameInOnly.chpl
	test/visibility/rename/renameMultiple.chpl
	test/visibility/rename/renameMultiple2.chpl