Save spectests .dat file to the performance description subdir
We've been getting sporadic failures to geneate the spectests for nightly
chapcs performance testing. This was because we were generating the file to
chapcs/spectests.dat instead of chapcs/<performancedescription>/spectests.dat.
This led to a race between all the chapcs perf runs, occasionally resulting in
one deleting the spectests.dat file, while another one was trying to use it.
This is hopefully the last instance of the perf playgrounds fighting with each
other (#3177, #3223)