Update jemalloc shim to use size_t for number of arenas
I think I was looking at an outdated man page when I though "opt.narenas" was
unsigned. Turns out it's size_t in 4.04. I discovered this when I added error
checking to the je_mallctl functions, which complained about being given the
wrong type.

I also update any comments referring to the man page to specify that it's the
4.0.4 man page.