Add jemalloc .good file for goodAllocSize
`chpl_mem_goodAllocSize(minSize)` returns the size the memory layer will
actually grab for a call to *alloc(). Since we're just using the standard API
for jemalloc right now, we just return minSize. When we switch to using the
extended API we'll be able to give a real answer using `nallocx()`