Parser changes to start implementing the ability to rename symbols in a 'use'
Created a union type to store either an element of an 'only' or 'except' list,
or a renaming pair of elements.  This was used to pass all of the list to
buildUseStmt.  buildUseStmt traverses and switches between the two options, but
doesn't do anything with the renamed symbols yet.

The compiler will build and three of the tests I wrote now pass.  Will verify
that 'only' and 'except' still work.