Add a mutex to our jemalloc chunk alloc hook
It looks like our chunk_alloc might not be called inside a lock like I thought
it was. This adds a simple mutex to protect getting the current pointer into
the shared heap. Note that the critical section (computing our current index
into the shared heap) will be very fast, so we shouldn't have much contention
on this lock (uncontended locks/unlocks happen in user-space and are fast)

This resolves the knucleotide regressions for gasnet-fast mentioned in #3252

BenH provided the following reduced knucleotide-like test case that allowed me
to track this down:

```chapel
proc calculate() {
  var freqDom : domain(uint);
  var freqs : [freqDom] int;
  for i in 1..100000 do freqs[i:uint] += i;
  return freqs;
}

proc main() {
  var ret = calculate();
  writeln(ret.size);
}
```

Without the lock this failed every ~3 runs, and with the lock it passed 10,000 trials