Resolve some coverity errors with the new UseStmt AST node.
I had misread the description of strlen and thought that it gave me a number
which included the string terminator, when it did not.  I also wasn't free'ing
a vector I had allocated in a couple of places, but the vector didn't need to
outlive its scope so I decided to just make it stack local instead of calling
new on it.