Cleanup jemalloc chunk_alloc function
Remove code that aligned size, which was used to calculate the updated heap
offset pointer.. I originally copied it from the tcmalloc shim, but there it
was used to set an `*actualSized` variable. Here we don't want it since we
align our base pointer, but don't need to align our size since jemalloc will
only request sizes that are a multiples of its chunk size.