Adding missing standard modules to the chpldoc documentation
In version 1.12 of 'chpldoc', it automatically crawled all
transitively 'use'd modules as it documented.  This meant that we
didn't have to list all modules/standard/*.chpl modules in our list of
files to chpldoc and so, apparently, we didn't.  While looking up some
documentation today, I realized that doc/master/ didn't have the
Time.chpl module documented and Lydia and Kyle correctly determined
that this change was the reason.  Here, I'm adding modules from
modules/standard that were included in 1.12 but missing in doc/master.

In passing, I'll note that virtually all of the modules/standard files
are now documented (Prefetch was a notable outlier) and that it would
be easier to maintain this if we could just use
modules/standard/*.chpl rather than explicitly listing modules to
document.  We're currently having a chat about the reasons for
neglecting that module.