Quiet mallocOutOfMemory for jemalloc with a shared heap
mallocOutOfMemory sets a virtual memory limit in order to test that we get a
sane error message on OOM. However, the limit it's setting is too low for
jemalloc when there's a shared heap. With a shared heap, we have to waste
whatever memory jemalloc got from the system before it was given our heap. This
is more than the current ulimit, so I'm just bumping it and the max heap size.

I just arbitrarily chose numbers that worked, and tested them out for a few
other configurations (gasnet-fast +dlmalloc, gasnet-everything, std,
--baseline, --fast) to make sure I didn't break them.