Partial implementation of renamed list in UseStmt and scopeResolve
Includes updating the constructor to take in a map of new_names to old_names,
updating UseStmt::verify and UseStmt::isPlainUse to ensure that the rename
list was treated like an extension of a limitation list, and adding a couple of
helper methods for determining if the name searched for was a rename and
getting the original based on it.  Also includes updating UseStmt::validateList
to check that the old_names are all valid symbols in that context, and to check
for the right name when the name being searched for is in the new_name list.

This also might fix a potential bug with naming conflicts (which I haven't
verified exists).