Don't skip looking for a name if it's a new_name in the rename list
Though I'd updated the name to look for in the module's symbolTable, I hadn't
updated UseStmt::skipSymbolSearch to take the rename list into account, so
we weren't even getting to the code needed to look for old_name in the
appropriate context.  This change fixes that, so now all but a few of the tests
I wrote pass.  Should probably write more tests.