Adjust tests for problems under valgrind
Removed writeRecordInOutProc.timeout and writeRecordOutProc.timeout
since 10s is an unreasonable valgrind compilation timeout
under valgrind (2 minutes is more like it).

Added a .future for ref-return-intent6.chpl. It is showing
a use-after-free in a function return pattern. This is
showing a problem with auto copies. It is probably related
to what Mike has been investigating.