Solve the transitivity issue with scopeResolve
In UseStmt::applyOuterUse, when traversing the 'only' lists of the outer or
inner use, also check if these identifiers are in the rename lists of the
other use, and if so, place the pairing there instead of in the normal only
list.  If there is a rename chain (i.e. Foo renames a to b, and Bar uses Foo
but renames b to c), cut out the inner link so that what is stored bridges that
gap (in the earlier example, what would be stored is a rename of a to c).

Also add trivial comment to description of field, and move a variable into a
more restrictive scope.