Suppress ISx perf timeouts for bradc-lnx and multi-locale gasnet
ISx is timing out on bradc-lnx and for multi-locale gasnet performance runs.
Since this benchmark is actively being worked on, I don't expect these
suppression to be around for long.

For gasnet-aries ISx takes ~1800 seconds. A lot of that time (800 seconds) is
in the exchange step, which we expect to get a lot faster with an optimization
Ben H. is working on.

I didn't time it on bradc-lnx, but it's probably close to the threshold since
chap03 only takes ~160 seconds. Although chap03 has 4GB of ram and bradc-lnx
only has 2GB so maybe we're swapping on bradc-lnx?

See JIRA issue 181 for more: https://chapel.atlassian.net/browse/CHAPEL-181