Add ability skip tests that require more locales than are available
This adds the ability to tell our testing system how many locales are available
to execute tests on and it will skip tests that require more.

To do this you can set `CHPL_TEST_NUM_LOCALES_AVAILABLE` and tests that require
more will be skipped with a "Warning" printed out and captured in the summary.

This is primarily motivated by our KNC test machine no longer have enough nodes
to run release/examples on. I'm also removing the KNC skipif for distributions
since it will no longer be needed