Un-futurize test_many_threads.
This test should work for tasking layer implementations that can switch
tasks on threads when a task blocks on a sync or single var, and should
not be expected to work for tasking layer implementations than cannot do
so.  (It may work in the latter case if the tasking layer can create
enough threads, but that's not dependable.)  In our case that means it
should work for qthreads and muxed, but not for fifo.  So, turn this
future into a regular test and skip it for fifo.  As part of this, fix
the .good so that it contains the output expected for the default config
setting of 'n' in the test.

We haven't tested this with massivethreads yet but we expect it to work
there as well, so allow that through the skipif also.