Change the active union field from an AST id to field number
Unions were storing an AST id and using that determine which field was
active. This ID is not stable across compiles, and works poorly with
generic unions (TBD if we want to allow generic unions though).

After this commit we now store the field number that is active. This
will be stable as long as the union's fields do not change (including
ordering).

Also removes PRIM_FIELD_ID_BY_NUM which was only used by unions to
support the old behavior.