Allow renaming of functions and find (type) constructors for renamed types
Since the name of the (type) constructor to look for was generated from the
type symbol already found, checking for a renamed version of either constructor
wouldn't help us.  It occurred to me that the easiest way to get around this
was to add the (type) constructor name to the relatedNames vector stored by the
UseStmt and be done with it.  This removed the need for the
matchedNameOrConstructorHelper method, so I simplified that implementation.

Verified that 'only' and 'except' weren't broken by this change.