Revert "Removing ISx performance timeouts"
[pointed out by @ronawho]

This reverts commit cf3e437208025abf30f876c6f6e49967c28d531f.

In a moment of foolishness, I thought of the default timeout as
being 500 seconds (which we're well below) rather than the 300
seconds it actually is.  In my defense, I'm told that I'm not
the only person who has this problem.  In any case, sorry
for the noise.

With the idxType change and Ben's bulk comm optimization, this test is
now right at the cusp of timing out on bradc-lnx.  While that makes
the timeouts in these files much higher than they need to be, for now
I'm taking the route of simplicity by reverting the change for now
while also looking for other ways of reducing the execution time.

The high overhead on bradc-lnx seems to be due predominantly to
creating and returning local arrays (which become a deep copy today)
on memory-contrained bradc-lnx.  I may prototype a version that passes
arrays in by reference for the sake of comparison.  This may also be a
useful comparison to make as we continue to wrestle with array return
semantics and implementation.