Add warnings/errors when identifiers are repeated in 'except' or 'only' lists
Catch when the user accidentally mentions the same symbol multiple times in an
'except' or 'only' list.  There are some cases where this isn't a catastrophe,
such as when the same symbol is renamed twice - in such a case, one could access
the symbol from either new name, but no conflicts are created so no harm is
done.  There are other cases where repeating a name is problematic, such as
when you've renamed a symbol to a name that you were already including.

Conflicts:
	test/visibility/rename/renameConflictInList.good