This is a variation of ISx that doesn't return local arrays
Our original ISx returned local arrays from each of the
procedures.  This variation declares the arrays at the
callsite and passes them in.  The thought was that this
would avoid unnecessary array copy overhead.  In a quick
check on my Mac, though, I'm not seeing that benefit.
Pushing this to my fork in order to try it on other
platforms.