Hacky fix for jemalloc w/ hugepage issues
I've been getting OOM issues when using jemalloc with hugepages loaded. It
turns out the reason is because we're generating some bogus code that's calling
chpl_memalign with an alignment of 0, but alignment must be a power of 2 at
least as large as sizeof(void*).

It's a bug that we're calling it like this, but for now I'm working around this
by just calling malloc if alignment is zero. Note that our other allocators
seem to work, so presumably they handle an alignment of zero, but jemalloc
chooses not to.