Remove dlmalloc source code and supporting infrastructure
This removes everything related to dlmalloc from our code base. We used to use
dlmalloc with gasnet segment large/fast, but as of #3287 we use jemalloc.

I basically searched the repo for all things related to dlmalloc and removed
them except for mentions in the CHANGES file. I then looked at the PR that
added jemalloc support (#3055) to make sure this I didn't miss
removing/updating any files.

This also updates the docs to reflect that jemalloc is used when comm=gasnet