Just limit the IPA level instead of throwing -O1 for CCE builds of jemalloc
IPA is the specific optimization that prevents jemalloc from building under
CCE. Instead of disabling most optimizations by throwing `-O1` just throw
`-hipa2` (`-hipa0 if `OPTIMIZE=0`.) The default is `-hipa3`  and that or
anything higher results in a build failure.

This allows us to build with a pretty high level of IPA and all other
optimizations enabled.