Fix unintentional naming conflict bug related to the 'except' list
By having the check for whether a use should be skipped based on its 'except'
or 'only' list at the same level as the check for if the use wasn't NULL in
scopeResolve, I was accidentally causing the remainder of module uses at that
scope to be skipped if a match for a symbol name had already been found.  This
meant that if a naming conflict would result from uses both before and after the
'except' or 'only' use, we would fail to detect the second one.