Fix DefaultRectangularArr.dsiSerialWrite's use of channel.writeBytes
The dsiSerialWrite function was passing in the base data pointer to
the writeBytes function. This would result in incorrect output if the
array was a sliced offset. The solution is to pass in a ddata shifted
based on the domain's lowest index.

This was exposed by 7021419c, which changed the result of the function
"isDataContiguous".